// api/clients/tmdb/tmdb-client.d.ts
export interface operations {
// ...
    "account-get-favorites": {
        parameters: {
            query?: {
                language?: string;
                page?: number;
                session_id?: string;
                sort_by?: "created_at.asc" | "created_at.desc";
            };
            header?: never;
            path: {
                account_id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        /**
                         * @default 0
                         * @example 1
                         */
                        page: number;
                        results?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /se5Hxz7PArQZOG3Nx2bpfOhLhtV.jpg */
                            backdrop_path?: string;
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 9806
                             */
                            id: number;
                            /** @example en */
                            original_language?: string;
                            /** @example The Incredibles */
                            original_title?: string;
                            /** @example Bob Parr has given up his superhero days to log in time as an insurance adjuster and raise his three children with his formerly heroic wife in suburbia. But when he receives a mysterious assignment, it's time to get back into costume. */
                            overview?: string;
                            /**
                             * @default 0
                             * @example 71.477
                             */
                            popularity: number;
                            /** @example /2LqaLgk4Z226KkgPJuiOQ58wvrm.jpg */
                            poster_path?: string;
                            /** @example 2004-10-27 */
                            release_date?: string;
                            /** @example The Incredibles */
                            title?: string;
                            /**
                             * @default true
                             * @example false
                             */
                            video: boolean;
                            /**
                             * @default 0
                             * @example 7.702
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 16162
                             */
                            vote_count: number;
                        }[];
                        /**
                         * @default 0
                         * @example 4
                         */
                        total_pages: number;
                        /**
                         * @default 0
                         * @example 80
                         */
                        total_results: number;
                    };
                };
            };
        };
    };
// ...
}